refactor: refactored the code for triangle
