+++ client/components/canvasTools/id.tsx
@@ -1,5 +1,6 @@
 export function generateId (object: any) {
-    let idArray = [object.left, object.top, object.width, object.height]
+    let randomNumber = Math.floor(Math.random()* 1000)
+    let idArray = [randomNumber, object.left, object.top, object.width, object.height]
     let idString = idArray.join("").split(".").join("")
     return idString
-}
+  }
