+++ client/components/panel.tsx
@@ -5,18 +5,17 @@ import { fabric } from 'fabric'
 import { clientSocket } from './fabricCanvas'
 import { Socket } from 'net'
 import { addText } from './canvasTools/text'
+import { addCircle } from './canvasTools/circle'
+import { addRectangle } from './canvasTools/rectangle'
+import { addTriangle } from './canvasTools/triangle'
 export const channelId: String = '/canvas'
 
 class Panel extends React.Component<PanelStateProps & PanelDispatchProps> {
 
   constructor(props) {
-    super(props)
+    super(props) 
     this.clearCanvas = this.clearCanvas.bind(this)
     this.handleClick = this.handleClick.bind(this)
-    this.generateId = this.generateId.bind(this)
-    this.addCircle = this.addCircle.bind(this)
-    this.addRectangle = this.addRectangle.bind(this)
-    this.addTriangle = this.addTriangle.bind(this)
   }
   componentDidMount() {
   }
@@ -29,65 +28,6 @@ class Panel extends React.Component<PanelStateProps & PanelDispatchProps> {
  async handleClick(action: string) {
    await this.props.updateTool(action)
 
-  }
-
-  generateId (object: any) {
-    let randomNumber = Math.floor(Math.random()* 1000)
-    let idArray = [randomNumber, object.left, object.top, object.width, object.height]
-    let idString = idArray.join("").split(".").join("")
-    return idString
-  }
-
-  addCircle() {
-    const newCircle = new fabric.Circle({
-      radius: 15,
-      left: 100,
-      top: 100, 
-      fill: this.props.color
-    });
-    newCircle["uid"] = this.generateId(newCircle)
-    let circleCommand = {
-      id: newCircle["uid"],
-      circleObject: newCircle
-    }
-    clientSocket.emit('draw-from-client', channelId, circleCommand)
-    this.props.canvasRef.add(newCircle)
-  }
-
-  addRectangle() {
-    const newRectangle = new fabric.Rect({
-      left: 100,
-      top: 100,
-      fill: this.props.color,
-      width: 20,
-      height: 20
-    })
-    newRectangle["uid"] = this.generateId(newRectangle)
-    let rectangleCommand = {
-      id: newRectangle["uid"],
-      rectangleObject: newRectangle
-    }
-    clientSocket.emit('draw-from-client', channelId, rectangleCommand)
-    this.props.canvasRef.add(newRectangle)
-  }
-
-  addTriangle() {
-    const newTriangle = new fabric.Triangle({
-      left: 50, 
-      top: 50,
-      width: 30, 
-      height: 30,
-      fill: this.props.color
-    })
-
-    newTriangle["uid"] = this.generateId(newTriangle) 
-    let triangleCommand = {
-      id: newTriangle["uid"],
-      triangleObject: newTriangle
-    }
-
-    clientSocket.emit('draw-from-client', channelId, triangleCommand)
-    this.props.canvasRef.add(newTriangle)
   }
 
   render() {
@@ -112,7 +52,7 @@ class Panel extends React.Component<PanelStateProps & PanelDispatchProps> {
         
         <button type="button" onClick={() => {
           this.handleClick('text')
-          addText(this.props.color, channelId, this.props.canvasRef)
+          addText(this.props.color, this.props.canvasRef)
           }}>Text</button>
 
         <button type="button" onClick={() => {
@@ -121,17 +61,17 @@ class Panel extends React.Component<PanelStateProps & PanelDispatchProps> {
 
         <button type="button" onClick={() => {
           this.handleClick('circle')
-          this.addCircle()
+          addCircle(this.props.color, this.props.canvasRef)
           }}>Circle</button>
 
         <button type="button" onClick={() => {
           this.handleClick('rectangle')
-          this.addRectangle()
+          addRectangle(this.props.color, this.props.canvasRef)
           }}>Rectangle</button>
 
         <button type="button" onClick={() => {
           this.handleClick('triangle')
-          this.addTriangle()
+          addTriangle(this.props.color, this.props.canvasRef)
           }}>Triangle</button>
